Put parameters at the bottom in README
The README is the first thing users will see on github and puppetforge. Newcomers will most likely want to know what this module does and how to use it. On most modules the parameter list is at the end of the file, so new users might thing this module is lacking examples and other basic documentation even though it exists at the bottom of the file.